requirements.txt
aiogram>=3.7,<4.0
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
numpy>=1.26.0
rapidfuzz>=3.9.0
transformers>=4.42.0
torch>=2.2.0
paddleocr>=2.8.1
opencv-python-headless>=4.10.0.84
python-telegram-bot>=21.0
